+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 - Met à jour la règle `@typescript-eslint/ban-ts-comment`: Un commentaire explicatif 
   est maintenant exigé lors de l'utilisation de `// @ts-expect-error` et `// @ts-ignore`.
 - Nouvelles règles:
-  - Les imports de type doivent toujouus être effectués via `import type ...`;
+  - Les imports de type doivent toujouus être effectués via `import type ...`.
+  - Interdit les séquences d'échappement \8 et \9 dans les chaînes.
+  - Empêche l'utilisation du chaînage optionnel dans les contextes ou `undefined` n'est pas permis.
+  - Les définitions d'objets avec membres non connus doivent être effectuées via `Record<..., ...>` (et non `{ [key: ...]: ... }`).
+  - Empêche l'utilisation d'expressions de type `void` dans des contextes déroutants.
+  - Empêche l'utilisation de `as` là ou `!` serait plus approprié.
 
 ## 1.0.0 (2020-08-11)
 - Première version "stable".